client-server/requirements.txt
paho-mqtt
flask
requests
urllib3
tinydb
schedule
py-healthcheck
redis